App/requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.1
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.20
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0